requirements.txt
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
pylint==2.3.1
Flask-Bcrypt==0.7.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-Script==2.0.6
Flask-WTF==0.14.2
Pillow==6.0.0
